extract common classes for server request decompressor
